refactor: move document-related utility functions to helper module
